* new-payment-methods:
  Add Payment interaction is now a push animation.
  da lang fixes
  Added Danish Localization
  revert
  Fixed product name
  Fixed allowing a user to check out without entering an address
  Hint view delay and shadow
  Fix plus and minus buttons
  New payment UX fixes
  Remove API key
  Fixed circle
  New payment method UX
  Added assets
  Made changes to the payment screen

# Conflicts:
#	Kite-SDK/KitePrintSDK.xcodeproj/project.pbxproj

+++ Kite-SDK/PSPrintSDK/OLImageDownloadDelegate.m
@@ -33,7 +33,7 @@
 
 - (void)URLSession:(NSURLSession *)session downloadTask:(NSURLSessionDownloadTask *)downloadTask didWriteData:(int64_t)bytesWritten totalBytesWritten:(int64_t)totalBytesWritten totalBytesExpectedToWrite:(int64_t)totalBytesExpectedToWrite{
     if (self.progressHandler){
-        self.progressHandler(downloadTask, totalBytesWritten, totalBytesExpectedToWrite);
+        self.progressHandler(downloadTask, (int)totalBytesWritten, (int)totalBytesExpectedToWrite);
     }
 }
